Fix compilation ambiguity in Solaris

+++ src/utils/utils.cpp
@@ -40,7 +40,7 @@ double kahan_sum(const SurrogateMatrix<double>& x)
  * https://math.stackexchange.com/questions/646117/how-to-find-a-function-mapping-matrix-indices
  */
 void s2d(const id_t id, const id_t nrow, id_t& i, id_t& j) {
-    j = nrow - 2 - static_cast<id_t>(sqrt(-8 * id + 4 * nrow * (nrow - 1) - 7) / 2 - 0.5);
+    j = nrow - 2 - static_cast<id_t>(sqrt(static_cast<double>(-8 * id + 4 * nrow * (nrow - 1) - 7)) / 2 - 0.5);
     i = id + j + 1 - nrow * (nrow - 1) / 2 + (nrow - j) * ((nrow - j) - 1) / 2;
 }
